requirements/base.txt
Collectfast==0.2.1
Django==1.7
Pillow==2.5.3
beautifulsoup4==4.3.2
boto==2.32.1
dj-database-url==0.3.0
dj-static==0.0.6
django-allauth==0.18.0
django-appconf==0.6
django-autoslug==1.7.2
django-avatar==2.0
django-braces==1.4.0
django-cache-url==0.8.0
django-configurations==0.8
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.1
django-floppyforms==1.2.0
django-heroku-memcacheify==0.5
django-model-utils==2.2
django-pylibmc-sasl==0.2.4
django-secure==1.0
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==2.4.2
gevent==1.0.1
greenlet==0.4.4
gunicorn==19.1.1
lxml==3.4.0
oauthlib==0.6.3
psycopg2==2.5.4
pylibmc==1.3.0
python-dateutil==2.2
python-openid==2.2.5
pytz==2014.7
requests==2.4.1
requests-oauthlib==0.4.1
sendgrid==0.5.1
sendgrid-django==1.0.3
six==1.8.0
smtpapi==0.1.2
sqlparse==0.1.11
static3==0.5.1
unicode-slugify==0.1.1
unittest2==0.5.1
urllib3==1.9
wsgiref==0.1.2
